+++ src/js/pages/index.js
@@ -1,9 +1,0 @@
-/**
- * Page Components Index.
- *
- * @package VmfaRulesEngine
- */
-
-export { OverviewPage } from './OverviewPage';
-export { DashboardPage } from './DashboardPage';
-export { ConfigurePage } from './ConfigurePage';